files merge
I've renamed the dashboard css and js files as cssdash and jsdash and also its index to indexdash so please kindly change it in the source code when linking any files with each other with their respectivr names.

Morover, i'll be finishing of the its UI and backend integration today itself, jist keep on pushin your commits too.